Refactor `ApplyEditAdapter.applyEdits`…
…for extra guarantees.

Each edit in a batch describes a buffer range and its replacement text. 
But they can't be applied in parallel. Each edit will change the buffer 
as it's applied, thus running the risk of making the original buffer 
ranges stale.

One way to “solve” this is to sort the edits by affected buffer range 
position and then apply them in reverse order, so that earlier edits 
likely won't affect later edits. But a better way is to apply markers to 
all the buffer ranges before any edits are made — so that we can follow 
the logical ranges around as they shift in response to edits.

We're still applying the edits in reverse order, but this gives us a 
belt-and-suspenders guarantee that would hold even if the language 
server is doing weird things — like suggesting edits that overlap one 
another.