Reformat AutocompleteAdapter comments…
…and apply a more targeted fix for the TypeScript autocompletion 
problem.

TypeScript autocompletions seem to fail when the suggestion's `text` is 
reassigned to be the same as `textEdit.newText`. This seems like a silly 
thing to do in the first place. Maybe there's a good reason for it, but 
the TypeScript server only works when that line is commented out.