Add `shouldIgnoreMessage` hook for linting
The goal here is to make it so that a consumer package can decide to 
disable linting selectively or altogether — perhaps by inspecting 
certain messages and their codes and comparing them to the user's 
configured preferences, or perhaps by inspecting the editor and finding 
that linting is disabled altogether for that kind of grammar.

This gives more granular control than all-or-nothing enabling/disabling, 
and it lets users change settings on consumer packages without requiring 
a window reload.