Allow an edit to start at a point past the end of an existing line
This seems to be needed for `typescript-language-server`. This is OK, because `TextEditor#setTextInBufferRange` and `TextBuffer#setTextInRange` seem to be just fine with the starting point of the old range being a point past the end of a line — it gets normalized to the end of the line.

This fixes the inability to execute “Add missing function declaration” in `pulsar-ide-typescript-alpha`.