Replace `zadeh` with Pulsar’s native fuzzy-matcher…
…but fall back to `fuzzaldrin` for safety.

`zadeh` doesn't have an Apple Silicon prebuild, so it causes headaches for some Mac users. Pulsar’s built-in fuzzy-matcher is fast and already there.

This would create a hard dependency on Pulsar >=v1.112 (I think) and would break this package's (theoretical) compatibility with Atom, so if that API isn't available, we fall back to `fuzzaldrin`. Slower, of course, but it's implemented in pure JS and (to be honest) `zadeh` was mostly overkill here — we do fuzzy-matching in _one_ place.