Add preliminary file-operation capability support…
…based on the proposed new `tree-view` service.

These operations have been listed as BLOCKED for ages because `tree-view` didn't provide hooks around user-initiated file operations. The proposed `tree-view` service would fix this.

If the service is present, we'll be able to consume it and report that we support these capabilities.

An LSP can do clever things with them — for instance, automatically adjust relative import paths when a file is moved to a different directory.